#3 - refactored iterator implementation
